requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
passlib
pyjwt
flask-cors
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.5
Pillow==9.3.0
requests==2.28.1
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13